move classes to better packages; remove FileUpdateEvent
